db: replace sqldump service with docker healtcheck
Abuse the healthcheck cmd by having it backup our database on a schedule.
This is cleaner than installing a separate cron/timer service in the image.
It also has less overhead than running a separate sqldump container/service.
However this is NOT what healtchecks were designed to do so this is a hack.
For now run the backup every 10 minutes and retry up to 3 times.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>